pyproject.toml
[project]
name = "campusx-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[google-genai]>=0.3.27",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.4",
    "langsmith>=0.4.13",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "langchain-tavily>=0.2.11",
    "langchain-community>=0.3.27",
]
